consolidating fixtures
