+++ src/tracksdata/graph/_test/test_subgraph.py
@@ -3,7 +3,6 @@ from collections.abc import Callable
 
 import polars as pl
 import pytest
-from test_graph_backends import graph_backend  # noqa: F401
 
 from tracksdata.constants import DEFAULT_ATTR_KEYS
 from tracksdata.graph._base_graph import BaseGraph
